requirements.txt
asgiref==3.5.2
Django==4.1.3
django-phone-field==1.8.1
gunicorn==20.1.0
Pillow==9.3.0
sqlparse==0.4.3
whitenoise==6.2.0
